backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "boto3>=1.42.16",
    "celery>=5.6.0",
    "django==6.0",
    "django-allauth==65.13.1",
    "django-celery-results==2.6.0",
    "django-cors-headers==4.9.0",
    "django-csp==4.0",
    "django-debug-toolbar==6.1.0",
    "django-environ==0.12.0",
    "django-filter==25.2",
    "django-storages==1.14.6",
    "django-stubs==5.2.8",
    "django-stubs-ext==5.2.8",
    "django-timezone-field==7.2.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "djangorestframework-stubs==3.16.6",
    "drf-spectacular>=0.29.0",
    "factory-boy>=3.3.3",
    "faker>=39.0.0",
    "ipython>=9.8.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "opentelemetry-instrumentation-django==0.59b0",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django==4.11.1",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "rich>=14.2.0",
    "ruff>=0.14.10",
    "sendgrid>=6.12.5",
    "sentry-sdk[django]>=2.48.0",
    "stripe>=14.1.0",
    "uuid>=1.30",
]
